+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntityQueryModelTest.java
@@ -23,16 +23,19 @@ import is.codion.common.model.condition.ConditionModel;
 import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.local.LocalEntityConnectionProvider;
+import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.OrderBy;
 import is.codion.framework.domain.entity.condition.Condition;
 import is.codion.framework.domain.entity.condition.Condition.Combination;
 import is.codion.framework.model.test.TestDomain;
+import is.codion.framework.model.test.TestDomain.Department;
 import is.codion.framework.model.test.TestDomain.Employee;
 import is.codion.framework.model.test.TestDomain.Job;
 
 import org.junit.jupiter.api.Test;
 
 import static java.util.Arrays.asList;
+import static java.util.Collections.singleton;
 import static org.junit.jupiter.api.Assertions.*;
 
 public final class DefaultEntityQueryModelTest {
@@ -48,7 +51,7 @@ public final class DefaultEntityQueryModelTest {
 	@Test
 	void condition() {
 		DefaultEntityQueryModel queryModel = new DefaultEntityQueryModel(new DefaultEntityTableConditionModel(Employee.TYPE,
-					CONNECTION_PROVIDER, new EntityConditionModelFactory(Employee.TYPE, CONNECTION_PROVIDER)));
+						CONNECTION_PROVIDER, new EntityConditionModelFactory(Employee.TYPE, CONNECTION_PROVIDER)));
 		EntityTableConditionModel conditionModel = queryModel.conditions();
 		assertFalse(conditionModel.get(Employee.DEPARTMENT_FK).enabled().get());
 		conditionModel.get(Employee.NAME).set().in("Scott", "John");
@@ -70,7 +73,7 @@ public final class DefaultEntityQueryModelTest {
 	@Test
 	void conditionChanged() {
 		DefaultEntityQueryModel queryModel = new DefaultEntityQueryModel(new DefaultEntityTableConditionModel(Employee.TYPE,
-					CONNECTION_PROVIDER, new EntityConditionModelFactory(Employee.TYPE, CONNECTION_PROVIDER)));
+						CONNECTION_PROVIDER, new EntityConditionModelFactory(Employee.TYPE, CONNECTION_PROVIDER)));
 
 		ConditionModel<String> nameCondition = queryModel.conditions().get(Employee.NAME);
 
@@ -84,10 +87,10 @@ public final class DefaultEntityQueryModelTest {
 		assertFalse(queryModel.conditionChanged().get());
 		queryModel.orderBy().clear();
 
-		queryModel.attributes().set(asList(Employee.NAME, Employee.JOB));
+		queryModel.attributes().included().set(asList(Employee.NAME, Employee.JOB));
 		nameCondition.operands().equal().set(null);
 		assertFalse(queryModel.conditionChanged().get());
-		queryModel.attributes().clear();
+		queryModel.attributes().included().clear();
 
 		nameCondition.operands().equal().set("Scott");
 		assertTrue(queryModel.conditionChanged().get());
@@ -105,7 +108,7 @@ public final class DefaultEntityQueryModelTest {
 		assertFalse(queryModel.conditionChanged().get());
 
 		queryModel = new DefaultEntityQueryModel(new DefaultEntityTableConditionModel(Job.TYPE,
-					CONNECTION_PROVIDER, new EntityConditionModelFactory(Job.TYPE, CONNECTION_PROVIDER)));
+						CONNECTION_PROVIDER, new EntityConditionModelFactory(Job.TYPE, CONNECTION_PROVIDER)));
 		assertFalse(queryModel.conditionChanged().get());
 		queryModel.having().set(Job.ADDITIONAL_HAVING::get);
 		assertTrue(queryModel.conditionChanged().get());
@@ -115,5 +118,46 @@ public final class DefaultEntityQueryModelTest {
 		assertTrue(queryModel.conditionChanged().get());
 		queryModel.get();
 		assertFalse(queryModel.conditionChanged().get());
+	}
+
+	@Test
+	void attributes() {
+		DefaultEntityQueryModel queryModel = new DefaultEntityQueryModel(new DefaultEntityTableConditionModel(Employee.TYPE,
+						CONNECTION_PROVIDER, new EntityConditionModelFactory(Employee.TYPE, CONNECTION_PROVIDER)));
+		queryModel.limit().set(1);
+		queryModel.attributes().included().set(asList(Employee.NAME, Employee.JOB));
+		Entity employee = queryModel.get().get(0);
+		assertTrue(employee.contains(Employee.NAME));
+		assertTrue(employee.contains(Employee.JOB));
+		assertFalse(employee.contains(Employee.MGR));
+		assertFalse(employee.contains(Employee.MGR_FK));
+
+		queryModel.attributes().included().clear();
+		queryModel.attributes().excluded().set(singleton(Employee.JOB));
+		employee = queryModel.get().get(0);
+		assertTrue(employee.contains(Employee.NAME));
+		assertFalse(employee.contains(Employee.JOB));
+		assertTrue(employee.contains(Employee.MGR));
+		assertTrue(employee.contains(Employee.MGR_FK));
+
+		queryModel.attributes().included().set(asList(Employee.NAME, Employee.JOB, Employee.MGR_FK));
+		employee = queryModel.get().get(0);
+		assertTrue(employee.contains(Employee.NAME));
+		assertFalse(employee.contains(Employee.JOB));
+		assertTrue(employee.contains(Employee.MGR));
+		assertTrue(employee.contains(Employee.MGR_FK));
+		assertFalse(employee.contains(Employee.COMMISSION));
+
+		queryModel.attributes().included().set(singleton(Employee.JOB));
+		employee = queryModel.get().get(0);
+		// All, since included and excluded cancel each other out
+		assertTrue(employee.contains(Employee.NAME));
+		assertTrue(employee.contains(Employee.JOB));
+		assertTrue(employee.contains(Employee.MGR));
+		assertTrue(employee.contains(Employee.MGR_FK));
+		assertTrue(employee.contains(Employee.COMMISSION));
+
+		assertThrows(IllegalArgumentException.class, () -> queryModel.attributes().included().set(singleton(Department.NAME)));
+		assertThrows(IllegalArgumentException.class, () -> queryModel.attributes().excluded().set(singleton(Department.NAME)));
 	}
 }